src/app/proveedores/proveedores.component.spec.ts
import { AuthService } from 'src/app/services/auth.service';
import { FirestoreService } from 'src/app/services/firestore.service';
import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
import { RouterTestingModule } from '@angular/router/testing';
import { IonicModule } from '@ionic/angular';

import { ProveedoresComponent } from './proveedores.component';
import { environment } from 'src/environments/environment';
import { AngularFireModule } from '@angular/fire/compat';

describe('ProveedoresComponent', () => {
  let component: ProveedoresComponent;
  let fixture: ComponentFixture<ProveedoresComponent>;

  beforeEach(() => {
    TestBed.configureTestingModule({
      declarations: [ ProveedoresComponent ],
      providers: [ FirestoreService, AuthService],
      imports: [RouterTestingModule, IonicModule.forRoot(), AngularFireModule.initializeApp(environment.firebaseConfig)]
    }).compileComponents();

    fixture = TestBed.createComponent(ProveedoresComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });

  it('Get Proveedores', () => {
    expect(component).toBeTruthy();
  });

  it('Display Proveedores', () => {
    expect(component).toBeTruthy();
  });

});
